pyproject.toml
[tool.poetry]
name = "google-group-members-list-python"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.51.0"
google-auth-oauthlib = "^0.5.2"
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
